Move board to own component
