+++ frontend/src/components/Board.css
@@ -1,0 +1,12 @@
+.board {
+  display: flex;
+  justify-content: center;
+  gap: 2rem;
+}
+
+.turnIndicator {
+  font-weight: bold;
+  text-align: "center";
+  margin-bottom: 1rem;
+  font-size: 1.2rem;
+}
